Rough in of adding line number to the annotation.
There's still a bunch of work to do for this, like adding a real db
field as we did for colors, to allow for filtering, but this at least
handles the part that determines the line number based on the selection.
We should be able to do something pretty similar to get the character,
too.